pyproject.toml
[project]
name = "reviewsentinel-playground"
version = "0.1.0"
description = "Testing playground for reviewsentinel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "openai>=1.64.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.0.1",
]
